Add partial for edit and create templates, make same path url for edit and delete method

+++ src/Controller/PinsController.php
@@ -116,12 +116,16 @@ class PinsController extends AbstractController
     }
 
     /**
-     * @Route("pins/{id<[0-9]*>}/delete",name="app_pin_delete", methods={"DELETE"})
+     * @Route("pins/{id<[0-9]*>}",name="app_pin_delete", methods={"DELETE"})
      */
-    public function delete(Pin $pin):Response
+    public function delete(Request $request,Pin $pin):Response
     {
-        $this->em->remove($pin);
-        $this->em->flush();
+
+        if($this->isCsrfTokenValid('123',$request->request->get('csrf_token'))){
+            $this->em->remove($pin);
+            $this->em->flush();
+        }
         return $this->redirectToRoute("app_home");
+
     }
 }
